ANTINSFWEMOJI & MASSEMOJI
this change added new error messages, prettified the code and added comments.

+++ src/functions/antiNsfwEmoji.js
@@ -6,20 +6,34 @@ async function antiNsfwEmoji(message, options = {}) {
     const embed = new Discord.MessageEmbed()
 
     try {
+
+        if (!message) throw new TypeError("\"message\" is not defined");
+        if (!options) throw new TypeError("\"message\" is not defined");
+        if(!options.embed && !options.text) throw new Error("\"options > text\" is not defined")
+
+        /////////////////////////
+        /// NSFW EMOJI TESTER ///
+        /////////////////////////
         const eggplantTest = /🍆/g.test(message.content);
         const peachTest = /🍑/g.test(message.content);
         const cherryTest = /🍒/g.test(message.content);
         const bananaTest = /🍌/g.test(message.content);
         const sweatDropsTest = /💦/g.test(message.content);
 
-        let NSFWEmojiUsage = false;
+        let NSFWEmojiUsage;
 
+        // if any of the tests returns true this changes "NSFEmojiUsage" variable to true 
         if (eggplantTest || peachTest || cherryTest || bananaTest || sweatDropsTest) NSFWEmojiUsage = true;
 
         if (NSFWEmojiUsage) {
-            message.delete()
-            let useEmbed = false;
 
+            message.delete() // Deletes the message
+
+            let useEmbed;
+
+            /////////////////////
+            /// EMBED CREATOR ///
+            /////////////////////
             if (options.embed) useEmbed = true;
             if (useEmbed && options.title) embed.setTitle(options.title)
             if (useEmbed && options.description) embed.setDescription(options.description)
@@ -30,7 +44,8 @@ async function antiNsfwEmoji(message, options = {}) {
             if (useEmbed && options.url) embed.setURL(options.url)
             if (useEmbed && options.thumbnail) embed.setThumbnail(options.thumbnail)
             if (useEmbed && options.image) embed.setImage(options.image)
-            message.channel.send(useEmbed ? embed : options.text)
+
+            message.channel.send(useEmbed ? embed : options.text) // sends embed or text to the message channel
         }
 
     } catch (err) {
